requirements.txt
requests
beautifulsoup4
mysql-connector-python
schedule
selenium
datetime
validators
telegram